business/requirements.txt
asgiref==3.5.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-ipware==4.0.2
django-microservices==0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
drf-spectacular-sidecar==2022.6.1
idna==3.3
inflection==0.5.1
jsonschema==4.6.0
Pillow==9.1.1
PyJWT==2.4.0
pyrsistent==0.18.1
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.28.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
